Fix UT after reorganization
